feat: se revisa que el usuario tenga un permiso, si no mandamos en 403

+++ app/Http/Controllers/CategoriaController.php
@@ -92,6 +92,8 @@ class CategoriaController extends Controller
      */
     public function destroy(Categoria $categoria)
     {
+        abort_unless(request()->user()?->can('eliminarCategorias'), 403, 'No autorizado');
+
         $categoria->delete();
         return response()->json(['message' => 'Categoria eliminada correctamente']);
     }
