+++ app/Http/Controllers/TiendaController.php
@@ -76,6 +76,8 @@ class TiendaController extends Controller
      */
     public function destroy(Tienda $tienda)
     {
+        abort_unless(request()->user()?->can('eliminarTiendas'), 403, 'No autorizado');
+
         $tienda->delete();
         return response()->json(['message' => 'Tienda eliminada correctamente']);
     }
